requestTranscript
