+++ apps/meteor/app/livechat/server/methods/requestTranscript.ts
@@ -5,7 +5,7 @@ import { Meteor } from 'meteor/meteor';
 
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { methodDeprecationLogger } from '../../../lib/server/lib/deprecationWarningLogger';
-import { Livechat } from '../lib/Livechat';
+import { Livechat } from '../lib/LivechatTyped';
 
 declare module '@rocket.chat/ui-contexts' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
@@ -31,6 +31,12 @@ Meteor.methods<ServerMethods>({
 		const user = await Users.findOneById(userId, {
 			projection: { _id: 1, username: 1, name: 1, utcOffset: 1 },
 		});
+
+		if (!user) {
+			throw new Meteor.Error('error-invalid-user', 'Invalid user', {
+				method: 'livechat:requestTranscript',
+			});
+		}
 
 		await Livechat.requestTranscript({ rid, email, subject, user });
